requirements.txt
# 核心依赖 - 系统基础功能所需
opencv-python>=4.7.0,<4.9.0
numpy>=1.23.5,<1.27.0
scikit-image>=0.19.3,<0.23.0
scikit-learn>=1.0.2,<1.4.0
pillow>=9.5.0,<10.2.0
pyyaml>=6.0,<7.0
streamlit>=1.24.0,<1.31.0
requests>=2.31.0,<3.0.0

# 深度学习依赖 - 用于模型训练和预测
# TensorFlow 2.15.0/2.15.1 仅兼容Python 3.9-3.11
tensorflow>=2.12.0,<2.16.0  # 根据Python版本自动选择兼容版本
ml-dtypes>=0.2.0  # MediaPipe兼容性需要

# 视频处理依赖
moviepy>=1.0.3,<1.1.0
imageio>=2.25.0,<3.0.0
imageio-ffmpeg>=0.4.8,<0.5.0

# 人脸检测和对齐依赖
dlib>=19.22.1,<19.25.0
face-alignment>=1.3.5,<1.5.0  # 1.3.5版本与模型修复脚本兼容
mtcnn>=0.1.1,<0.2.0
mediapipe>=0.10.3,<0.11.0  # 0.10.8版本与TensorFlow兼容性最佳

# 文件和数据处理
h5py>=3.8.0,<4.0.0  # 用于BiSeNet模型占位文件
bz2file>=0.98  # 用于dlib模型解压

# 其他实用工具
tqdm>=4.65.0,<5.0.0
matplotlib>=3.7.1,<3.9.0
pandas>=2.0.0,<2.2.0

# 系统依赖说明(不需要pip安装) - 根据环境确保以下依赖已安装
# CUDA >= 11.2 (用于GPU加速, TensorFlow 2.12.0-2.15.0兼容)
# cuDNN >= 8.1 (用于GPU加速)
# FFmpeg (用于视频处理)

# 可选和兼容性说明:
# - Python 3.8-3.11 兼容
# - 对于Python 3.8，建议使用TensorFlow 2.12.0-2.14.0
# - 对于Python 3.9-3.11，建议使用TensorFlow 2.15.0/2.15.1
# - 如果使用GPU加速，请确认CUDA和cuDNN兼容性
